+++ changelog.md
@@ -23,6 +23,7 @@ slide:
 - `footer(text: string, fontSize=20, opacity=0.6, rawHtml=false)` template to a footer to every slide. The input is a string with markdown, unless the argument `rawHtml=true` in which case it is treated as raw HTML.
   - The footer will be shown on all slides except those with image-, video- or iframe-backgrounds.
 - Support for making lists that incrementally reveal the list items are supported using the templates `orderedList`, `unorderedList` and `listItem`.
+- `addStyle` allows you to add CSS to the document.
 - Experimental features - the API for these are not yet stable and might change in the future:
   - `fragmentThen` - Allows the construction of fragments like `growThenShrink` by `fragmentThen(grows, shrinks)`. The second animations happens at the same time as the next animation, reducing the number of clicks needed.
   - `align` template to simply align its content like this: `align("left"):`